+++ src/Services/Main/Service/EventType.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Main\Service;
+
+/**
+ * Event handler delivery type used by event.bind / event.unbind.
+ *
+ * @see https://apidocs.bitrix24.com/api-reference/events/event-bind.html
+ */
+enum EventType: string
+{
+    case online = 'online';
+    case offline = 'offline';
+}